Webpage: corrected date of 0.14 release
In the main commit of Gnuastro 0.14, I also forgot to update the date of
the release on the webpage! This (as well as the previous commit, where I
forgot another item in the checklist) happened because I went through half
of the release checklist last night, but had to stop in the middle due to a
personal thing and continued after that in the morning!

It has been corrected with this commit.