Options: FITS/TXT file values specified in --help and book
Until now, we were simply using an 'STR' as a place-holder for the value
for options that accept strings. But this didn't differentiate between
strings like file names, or other strings like program-specific names, or
HDU names and etc. This was could be confusing for the users, but also make
it hard to enable the shell's auto-complete features (to see when the
auto-complete should show FITS files, or when it shouldn't).

With this commit, as a first step, any option that takes a file name as
input will now show 'FITS' or 'FITS/TXT' as the place holder for its value
in the output of '--help' or in the book.