New feature: Auto Complete (WIP)
Until now, the user did not have the convenience of using auto
completion in Gnuastro. With this commit, user can push the <TAB>
character in the middle of typing and select the appropriate option
or command.

This commit is a work in progress (hence the WIP in the subject). For
now, the feature is only available in `astcosmiccal`. But we are
planning to expand it to all Gnuastro subprograms in near future.