Completion: Suggest FITS files in relevant options
Until now, the completion did not suggest `FITS` files for relevant
options. With this commit when user types in the following command:
`astconvolve --kernel <TAB><TAB>`, suggestion includes only the files
present in the current directory that have the `FITS` extension.

It is worth to note that the `FITS` extension is set to be case
insensitive.

Also, now the suggestions are architectured differently. In more
details, suggestions are added to the `COMP_REPLY` array using the `+=`
operator. This is to prevent confusions and incomplete suggestions for
situations that the word being completed can satifsy two or more
conditions.