Library (fits.h): wrapper over WCSLIB bug (present from 2015)
Until now, WCSLIB would wrongly write the CDELT values as '0.0' for some
types of data-cubes (where the scales of the CDELT values in each dimension
was large, but not large enough to warrant a different printing
format). After contacting Mark Calabretta (author of WCSLIB), we discovered
that this bug bug has been present in WCSLIB since its version 5.9
(released on 2015/07/21). For more on the precise cause of the function see
the comments above the new 'fits_bug_wrapper_cdelt_zero' function.

Mark did mention that it will be fixed in the next release of
WCSLIB. However, it will take a long time for many users (actually their
package managers) to update. We therefore need to write a wrapper to fix
the situation when it happens.

With this commit, a bug-fix-wrapper function has been added just after
writing the WCS keywords into the FITS file: upon writing any of the CDELT
keyword lines into the FITS pointer, it will check if the written value,
and the value inside the 'wcsprm' structure differ by more than 10
decimals. When they do differ, by more than this, this function will use
CFITSIO to update the keyword value in the key header.

The operational footprint of this check is very small (only being activated
on datasets that have more than 2 dimensions, and only doing the check on
'CDELT' keywords, not the others).

Also, since this is the first change in Gnuastro's library after the
release of Gnuastro 0.14, I have incremented the Gnuastro library Soname
version and added the top-level lines in the NEWS file for the next
release.