Library (fits.h): date to seconds sets NULL pointer when no sub-second
Until now, when there was no sub-second value in the input date string, the
'gal_fits_key_date_to_seconds' function wouldn't touch the sub-second
string or value. This would require the user to initialize these pointers
before calling this function (which would not always happen!). This would
cause a problem in their checks afterwards. This actually happened in the
Fits program that called this function!

With this commit, when there is no sub-second in the input date, the
'gal_fits_key_date_to_seconds' function will put a NULL pointer in the
expected string and a NaN value in the expected sub-second value.

Also, just to avoid compiler warnings in the Fits program, I initialized
the sub-second string in the Fits program to NULL before calling
'gal_fits_key_date_to_seconds'.