Installation: LDADD found at configure time always used
Until now, the 'CONFIG_LDADD' linker options that are gathered during the
'./configure' step were only added when RPATH was not enabled at
configuration time. The basic idea was that when RPATH was enabled, the
Gnuastro library already linked with the dependency libraries, so there was
no need. However, when Gnuastro was bootstrapped on some GNU/Linux systems
the RPATH test didn't operate as expected and would cause a linking
problem.

With this commit, we directly use the 'CONFIG_LDADD' variable
unconditionally in the 'Makefile.am' of all the programs. After a test on
my own system, the final build command that was printed on the terminal was
identical between the two and made no difference. It also fixed the problem
on the computer of Francois (who initially reported this problem)

This bug was fixed with the help of Francois Ochsenbein and was previously
also reported by Sachin Kumar Singh.