Bootstrapping: shell check to see if tikz directory exists
Until now during the bootstrapping (where we build the figures used in the
paper), we were using 'mkdir -p'. But the '-p' option to 'mkdir' is not
portable and primarily exists in GNU Coreutils.

With this commit 'tikz' is now a Make order-only target, so we don't need
the '-p' option to 'mkdir' any more. This is because 'mkdir' is only called
when it doesn't exist.