Table: the --noblank option can now work on all columns
Until now, if you wanted to use the '--noblank' option on all the columns
of a table, you had to specify all their names! This was frustrating when
the table has many columns that should be checked, and could be buggy
(making a typo in the column names for example).

With this commit, this option now accepts a unique '_all' value. In this
mode, the user just has to give '--noblank=_all' and all the columns will
be checked.