astscript-radial-profile: simplifying the script
Until this commit, the radial profile script evolved in the direction of
having more and more options within it. However, this goes in the
opposite direction of the minimalistic and modularity principle. Tasks
such as the binning of the radial profile, the modification of the
columns of the final output, or the detection of signal to measure high
level values do not have to be computed in this script. The aim of this
script is clear: compute measurements over radial apertures of a given
object from a specific image.

With this commit, all of these not extrictly necessary tasks have been
removed from the script. Now, it only computes the radial profile
values. In addition to this, many other modifications have been done.
The most important one is the use of long-name variable inside the code.
This allows much faster identification of bugs/problems and the solution
of them. For example, instead of using the variable `x' for the x center
of the object, along with the entire script, now the `xcenter' variable
name is used. The same happen with the rest of variables.