Book: simplifying documentation of astscript-radial-profile
Until this commit, the documentation of the radial profile script were
not updated properly according with the modification of the previous
commits. So, the documentation was not describing the actual script.

With this commit, the description of the radial profile script and its
options have been modified in order to consider the latest
modifications. In addition to that, very minor modifications in the
actual code of the script was done. The main one is to not use the
program `realpath' when considering the temporal directory.