Tabel: Completion, examine for fits availability
Until now, after the user entered either commands '--information' or
'--wcsfile' the autocomplete would suggest a fits file name. This is a
bad idea when the user has already defined the file.

With this commit, we will check if the user has specified the file
already, if so, we will carry on with suggesting all other options at
hand. On the other hand, we will suggest a fits file living in current
directory.

Still, if there were no fits files in the current directory,
autocompletion will not suggest anything. This might signal the user
that something is wrong.

Also, I elaborated on the '--column' option's comments.