Table: Completion, fix 'wcshdu'
Until now, the 'wcshdu' option would work only without the equal sign.
With this commit, the 'compgen' will cease to filter out the words that
do not match the current word. This is because when the cursor is just
after the equal sign '=', it will be considered as the current word. And
none of the options contain '='. Hence, all of them will be removed from
suggestions.

Also, a short description of the 'wcshdu' command is added.