Table: two new options to limit rows: --rowlimit and --rowrandom
Until now, the only position-based row-selection options available to a
user where the '--head' and '--tail' options. But in some scenarios, we
want a certain range of rows based on their position in the full table (for
example rows 100 to 200). In other cases, to get a feeling of the dataset,
or for statistical tests, we want a random set of the rows. Until now, it
was not easy to extract these.

With this commit, the new '--rowlimit' option enables selecting a
contiguous range of rows anywhere within the table and the '--rowrandom'
enables selecting a random sub-set of the rows (after all value-based
selections have been applied).