Table: correct mistyped name for --rowrandom
This option was just added to table a couple of commits ago, but the actual
option name (and its description in '--help') was wrongly entered as
'--randomrows'.

With this commit, the option name has been corrected to have the same name
as the description in the book, and the relevant variable name within the
Table program's source.