Table: not using full permutation in row selection
Until now, when we wanted to select random rows, or generally select rows
by their values (for example with '--range'), we would permute the full
output table to bring the desired rows to the top. In other words, even the
rows that we didn't care about were being moved around (resulting in wasted
time/CPU).

With this commit, instead of using full permutations, we are just moving
the desired rows to the top of each column, and leaving the rest of the
table untouched. This greatly speeds up the selection of the desired
rows.

A new internal function called 'table_bring_to_top' has been defined to
easily use this feature in both the 'table_select_by_value' and
'table_random_rows'.