Table: Completion, bash4+ long option no space
Until now, we used the 'nospace' option of 'complete' to handle
suggestions for options which had an '=' inside them. With this commit,
if the command being completed contains '=', no spaces will be appended.

This only works in bash4+ for '[[' and 'compopt' are used within this
approach. Let's just keep this here and look for a workaround in future
commits.

TODO: Work on compatibility for bash older than version 4.