Table: Completion, custom compgen
Until now, the completions used the built-in 'compgen' command. But
'compgen' did not handle '-' inside a single suggestion such as
'QUERY-CONFIG' for '--wcshdu' option. With this commit, the
'_gnuastro_autocomplete_compgen' is introduced. The behavior is
completely predictable.

Also, the global variables are 'unset' when they are not needed.

Lastly, the '=' sign is handled well in this version. It is removed if
it is counted as the current word, and it is ignored if it was counted
as the previous word. This way we have a way of knowing which long
option is being completed. Furthermore, this prevents some complications
while we are filtering out the suggestions.