Table: Completion, add options, improve get name
Until now, the '_gnuastro_autocomplete_get_fits_name' function echoed
the first fits file the user specified. This breaks the behavior for
options such as '--catcolumns' because there might be several fits files
on the command line.

Nevertheless, everytime the user asks for the column names that belongs
to the last fits file specified. Hence, the function will iterate
through all words in the commandline, check them one by one and store
them as a fits file if the name is valid, and then continue the loop
until the last word. This way, we are making sure that the stored fits
file is always the lates file specified.