Bootstrapping: separate target for building all necessary images
The 'doc/plotsrc/Makefile' is in charge of building the figures in the
PDF/HTML manual. However, its main rule (to build 'doc/plotsrc/all.pdf')
didn't take into account that when a command fails in standard Make, the
rest of the commands in the rule continue without problem!

On the other hand, the main target in 'doc/Makefile' to build the figures
(that would trigger calling 'doc/plotsrc/Makefile') was one of the image
'.txt' files that was created relatively early.

As a result of both the issues above combined: when building of one of the
figures failed during bootstrapping, the error was hidden in the middle of
the output and only show itself when 'make pdf' was ran. Confusing us on
why the image is not built.

With this commit, the main target of 'doc/plotsrc/Makefile' is no longer a
file that is actually used. But a specific file called 'done.txt' which is
made in the 'doc/gnuastro-figures' directory, only when all the images are
built. Also, the recipe of this target is all effectively a single shell
line (with many individual commands, separated by '&&'). So if any of the
commands fail, the final target will never be built and nothing extra will
be printed, helping find the cause of the problem much more easily.

This issue was reported by Francois Ochsenbein.