Webpage: new coming soon page and SVG animation in checklist
Until now, for each release we would just suffice to the NEWS file. But to
help new users see the power of new features, for Gnuastro 0.15, we'll try
putting an SVG terminal-grab animation to hopefully get new users more
interested.

With this commit, the series of steps to do make an animation have been
added in the release checklist and a new 'doc/coming-soon.html' webpage has
been started (not yet complete) to act as a template for each new release.