Book: Discussion on Java added in the section on Why C
Until now, in the "Why C?" section of the "Developing" chapter of the book
we had primarily focused on comparing with C++ and Python. However, Java is
also extensively used in some parts of the astronomical community (for
example in some implementations of the Virtual Observatory standards like
TopCat, or in the Gaia project). It is therefore necessary to add a
discussion on Java here also.

With this commit, a paragraph on Java has been added discussing Java's
unique JVM and its effects, just after discussing Python. But before
generalizing the discussion (on the high cost of keeping up to date with an
evolving language).

Also, since "François" contains the cedilla accent over the 'c', an entry
in the 'doc/gendocs' script has been added to convert it to Texinfo
format. So it displays properly in the PDF, HTML, info or plain-text
outputs. Recall that 'doc/gendocs' creates the list of authors at the start
of the manual (that go in the second page of the PDF).

This commit has been edited by Mohammad Akhlaghi.