Installation: --disable-shared links statically with dependencies also
Until now, when '--disable-shared' was given, only the main Gnuastro
library was statically linked to the programs. All the other dependency
libraries would be linked dynamically, even if a static version of that
library also existed. This would make it hard to move the built binary to
another system (which may have different versions of the shared libraries).

The main problem was that the 'AC_LIB_HAVE_LINKFLAGS' Autoconf macro
(provided by Gnulib), ignores the '--disable-shared' option. As a result,
if a shared library existed, it would prefer that and return the shared
library instead of the static library.

With this commit, the issue has been fixed with a hack in the GNU
Portability Library (or Gnulib). The hack is implemented as a
'bootstrap_post_import_hook' in Gnuastro's 'bootstrap.conf'. In this hook,
a step was added to only check for shared libraries if the 'enable_shared'
variable has a value of 'yes'. With this hack, when '--disable-shared' is
given, the the returned value of 'AC_LIB_HAVE_LINKFLAGS' is the absolute
path of the static library (if it exists). If no static library exists on
the system, then a shared library is still found and returned (with
'-lNAME', where 'NAME' is the name of the library).

Therefore when '--disable-shared' is given if any dependency has a static
library, it will be statically linked into Gnuastro's programs (making them
portable).

This issue was found in a discussion with Sylvain Mottet.

In the process three other issues have been found and addressed:

 - When building in '--disable-shared' mode (as I commonly do personally
   during development!), the BuildProgram step in 'make check' failed with
   the new features! The cause was that the 'lib/libgnuastro.la' file
   doesn't include static libraries in its 'dependency_libs' variable. It
   also prints a bad warning that this type of "shared" library isn't
   portable (but its not a shared library, so I am ignoring it for now). To
   fix this, just before the 'make' step finishes, we will re-write the
   value of 'dependency_libs' in 'lib/libgnuastro.la' to be the full set of
   libraries (including static) that have been found during the configure
   phase. This fixed the crash during 'make check'.

 - While doing this test in an environment where all dependencies had
   static libraries, I noticed that CFITSIO can also depend on the BZ2
   library if it was present when CFITSIO was built. So we now also check
   for the BZ2 library before CFITSIO.

 - When 'AC_LIB_HAVE_LINKFLAGS' goes searching for static libraries,
   instead of '-lpthread', it can return 'libpthread.a'. However, at least
   in Gnuastro's static build on a GNU/Linux system, the POSIX threads
   library (which is part of the host's C library) will crash (as described
   in 'configure.ac'). So with this commit, a step has been added in the
   final stage of setting the library linking list that will replace any
   string that ends with 'libpthread.a' with '-lpthread'.