Installation: check for the C math library before all others
In the previous commit (which corrected the finding of linking flags at
configure time), I had removed the check for the C Math library because the
'AC_LIB_HAVE_LINKFLAGS' would add '-lm' in the check for GSL. However, on
an Ubuntu system, we just discovered a crash because of not being able to
link with the Math library.

With this commit the check for the Math library and the addition of its
flag has been re-introduced.

On a minor note in the build on the Ubuntu, we also found a simple compiler
warning about the possibility of being used uninitialized (which would
never happen). But to supress the warning, the variable (a pointer to a
function) has been initialized to NULL.

This issue was reported by Zahra Sharbaf.