astscript-radial-profile: fixed bug when creating the temporal directory
Until this commit, the conditional for creating the temporal directory
(where all the necessary files for constructing the radial profile are
kept) was not the correct one. With this commit, this has been fixed. In
addition to this, also minor corrections to some parts of the comments
have been added.