Fits & Table: new --keyvalue (Fits) and date-to-sec operator (Table)
Until now, the Fits program had no easy value to just view the values of
certain keywords. You had to get the list of all keywords, then use 'grep'
or 'awk' to extract the ones you want. This was inconvenient in two
aspects: extracting string values (which have a single-quote around their
values), or the need to select a sub-set of many FITS files.

With this commit a new '--keyvalue' option has been added to the FITS
program to address this issue and greatly simplify the inspection of many
files. This option will only read the requested keywords from any number of
input FITS files and write the results in a table format (that can easily
be pipped onto Table for example for selecting a sub-set of the FITS images
based on their keyword values.

Having done this, another useful feature was implemented: the Table
program's Column Arithmetic now has a new operator called
'date-to-sec'. With it, you can easily convert the standard FITS dates
(found in the 'DATE' or 'DATE-OBS' keywords of FITS files).

The following minor changes were also made in the process:

 - The FITS program can now take any number of input FITS file
   arguments. So the name of the input is now put in a list of strings
   instead of a simple string. However, for now, this is only used with the
   '--keyvalue' option. For the others, only the variable with the input
   filename has been updated.

 - In the blank library, for string types, we now also count NULL pointers
   (for each string) as blank. Also the 'gal_blank_number' function can now
   also work on string datasets (until now it only used the Statistics
   library, which was limited to numerical data types).

 - The 'gal_fits_key_date_to_seconds' function will return a blank output
   when the input is a blank string.

 - The 'gal_fits_key_read_from_ptr' function can now find the best value
   for a given keyword value when a type isn't specified.

I had been planning to add this feature for some time, but the final
motivation came after a discussion with Alberto Madrigal.