Installation: using proper linking flags for shared libraries
Three commits ago (commit e723937d: Installation: --disable-shared links
statically with dependencies also), I changed the output variable of
'AC_LIB_HAVE_LINKFLAGS' that is used to build the programs (from
'LTLIB<NAME>' to 'LIB<NAME>' when setting 'LDADD'). This was done primarily
focused on doing a static build of Gnuastro (configuring Gnuastro with
'--disable-shared').

However, I didn't notice that as a result of this change, linking problems
will pop-up in the default, shared-library mode, linking of Gnuastro. This
was a complex issue to understand, partly because of a not-well-enough
documentation of 'AC_LIB_HAVE_LINKFLAGS' in Gnulib. After a lot of trials
and comparison with previously good commits, I found out that 'LIB<NAME>'
should only be used when building static programs. When building shared
programs we should use 'LTLIB<NAME>'.

With this commit, this issue has been addressed and an extensive
description of the outputs of 'AC_LIB_HAVE_LINKFLAGS' is given in
'configure.ac' just before its first use.

This issue was found with the help of Alberto Madrigal.