Book: added another useful example for --keyvalue
Recently the '--keyvalue' option was added to the Fits program. However,
all the examples had a simple '*.fits' which is only good for the files in
the current directory.

With this commit, a new paragraph has been added to the description of this
option with an example of using the 'find' program for searching all FITS
files in the sub-directories of a certain directory.