Table: the set- operator now availabe in column arithmetic
The 'set-' operator has been very useful in the Arithmetic
program. However, until now this operator wasn't available in Table's
column arithmetic.

With this commit, the core functions that done the job for the Arithmetic
program have been moved into the Gnuastro library and thus the same
function is now being used in both the Arithmetic program and Table's
column arithmetic. In order to do this, some small polishing was done in
both the Arithmetic program and Table (the parts related to column
arithmetic).

In the process two minor issues have been found and corrected:

 - The error messages in the arithmetic library that reported the operator
   name didn't have any single-quotes around them.

 - The "multiply" operator was being mistakenly reported as '*' instead of
   'x'.

 - In the Fits program, when correcting the types of the columns as the
   data was being read, a typo had caused mistakenly using the 'key'
   variable instead of the correct 'write' variable.