Library (arithmetic.h): types properly set for integer operators
Until now when any of the integer-only operators (like modulo, or the
bitwise operators) was given operands that weren't unsigned 8-bit integers,
the output type was mistakenly set to unsigned 8-bit integers. Therefore
creating a segmentation fault.

Having a look into the source, I recognized that the issue is due to the
function that determines the type of the output. In that function (which
hasn't been touched for a VERY LONG TIME: before the problematic operators
were introduced) I was assuming that the binary operators are plus, minus,
multiply and divide as well as the conditional operators. Since the
conditional operators only unsigned 8-bit integers, I was returning this
type for anything that is not plus, minus, multiply or divide.

With this commit, the issue was fixed by reversing the check: if
conditional operators are requested, the output will be unsigned 8-bit
integers, otherwise, it will be the larger of the two input types.

In the process, I also discovered that when reporting errors on conditional
operators, I was mistakenly using the conditional operator signs (like '<'
for 'le', or less than). This has also been fixed with this commit.

This fixes bug #60082.