sort-by-night: dramatic improvement in speed with new features
Until now, astscript-sort-by-night primarily used the shell's 'awk' or
'sort' programs to almost all its operations. As a result, it wasn't too
efficient.

With this commit, the new features that have been recently added to the
Fits and Table programs are now used for the computational components of
the job: with the new '--keyvalue' option of the Fits program, the keyword
values are read from all the inputs in one command and that is piped to
Table which does all the calculations using the new 'date-to-sec' and
'set-' operators. Finally, Table is again used to sort the input at once
and at the start, and to select the proper night.

This resulted in a major improvement in the running speed of this script
(as mentioned in the NEWS file, from from 19 seconds to 0.42 seconds for
about 650 FITS files used in the test!).