Table: Completion, replace double negative
Until now, complete had the 'nospace' option by default, and it would
turn it off whenever there were no '=' signs in the 'COMPREPLY'. With
this commit, the 'nospace' option is off by default. It is only turned
on when there is a '=' sign in the 'COMPREPLY'. The logic might be a bit
more clear now.