Book: added discussion on converting pixels to microJanskys
In many cases, it is useful to convert an image's pixel values to
microJanskies (to compare pixel values between various instruments or
filters).

With this commit, a discussion has been added in the "Brightness, Flux,
Magnitude and Surface brightness" section of the book.

This was done after a discussion with Ignacio Trujillo, Raul Infante-Sainz
and Samane Raji.