New installed script to generate SAO DS9 region file from table
Generating an SAO DS9 region file from the output of Gnuastro's Table
program is very easy. It is also a very common job when analyzing the
outputs of object selection in a study. However, such a script requires
multiple multiple lines that users have to maintain themselves.

Because this is a very commonly required job during an astronomical
analysis, to simplify the work an Installed script has been added to
Gnuastro for this job with some basic operations.