make-ds9-reg: Added option to give each region a name
Until now, we made a region for each object with a defined coordinate and
radius but each region doesn't have any name. However, in some cases we
need to know the name of objects to recognize them

With this commit, a new and optional '--namecol' option has been added to
this installed script. When given, it will print the given column's value
over each region.