Library (units.h): function to convert counts to Janskys
Until now there was no easy way to convert counts to Janskys. An AWK
command was given in the section on Brightness and magnitudes, but it is
annoying to go and copy-paste it into a command!

With this command, a new function has been added to the units library:
'gal_units_counts_to_jy'. With this function in place, an operator in the
Arithmetic library has also been defined for this task and through the
library, it is possible to do this conversion on images or tables directly
on the command-line.