astscript-radial-profile: new --oversample option
Until now we could only use the raw resolution of the input. However,
except for the very central pixels, we are measuring values for each radial
distance over MANY pixels. Therefore, we can increase the measurement
resolution.

With this commit, a new '--oversample' option has been added to the
script. When this option is used, Warp will over-scale the input to finer
pixels (and correct the center and maximum radius). The over-sampled image
is then given to MakeProfiles and MakeCatalog. Finally, the first column
(containing the radius) is divided by the oversampling factor to give radii
in the original resolution (sub-pixel).

Besides this work some other important changes have been made with this
commit:

 - The old '--xcenter' and '--ycenter' options have been merged into a
   single '--center' (like Crop). The user give comma-separated values, and
   we separate them internally.

 - A sanity check was added to make sure that only 'img' or 'wcs' are given
   to '--mode'.

 - In the default name (when '--output' isn't given), there 'rprofile' is
   changed to 'radial_profile'.