Library (arithmetic.h): multi-operand operators free next pointer
Until now, the multi-operand operators (that take lists as input, like
'mean', 'median' and etc: those that take a number for the number of inputs
and operate on all the pixels) wouldn't set the 'next' pointer of their
output to NULL. They would just free all the other list elements.

As a result, when a future function/operator wanted to look into the
datasets, it would think it is a list (because the 'next' element wasn't
NULL), but the space pointed by 'next' was freed already! This would create
a segmentation fault!

With this commit, once the multi-operand operation is finished and it is
busy freeing the input list, when it hits the output dataset, it will set
its 'next' pointer to NULL. Thus fixing the problem.

In the process, I also discovered that the 'pop_number_of_operands'
function isn't freeing the popped dataset within the 'SET_NUM_OP' macro!
That is also fixed here.

This fixes bug #60121.