astscript-radial-profile: minor bugs in -v and -c options corrected
In the previous commits the two '--center' and '--oversample' options were
added or modified with the aim of make the script more general and easy to
use. However, there were a couple of mistakes (probably due to a copy &
paste of lines) and the short options did not correspond to the long
version of the agruments. Also, the documentation of the old '--xcenter'
and '--ycenter' still remained in the book (although they had been merged
into a single '--center' option with two values).

With this commit they have all been addressed.