Bash completion: renamed fixed file to completion.bash
Until now, the fixed file for shell completion within the source of each
program was called 'completion.sh'. This could hide the fact that this is
only for Bash, and make things complicated if we later want to add
completions for other shells.

With this commit, to be more explicit that this completion is only for GNU
Bash, the fixed file in the source of all programs is renamed to be
'completion.bash'. Also, an item for it was added in the "Program source"
section of the Developing chapter and a short intro into these files was
added in the "Auto-completion in Gnuastro" section.

Also, I noticed that the "Program source" section still didn't follow our
relatively recent (2 year old) convention of having one line per
sentence. So that has also been implemented for the full text of this
section.