bin/table/completion.bash: improvements to find good table name
I have tried to go over the current implementation of the completion script
for Bash. I succeeded in doing checks to find the last table file on the
command-line (even if it is a plain-text table). However, I hit a problem
when I wanted to implement a feature for the '--information' option, which
is described in a comment above the check (starting with 'PROBLEM').

Pedram, can you look into this problem and find a fix? This will be a
useful solution if we can add this feature (dynamically printing warnings
to the user).

Here is a summary of the points I changed:

 - All the global variables have a '_gnuastro' prefix in their name
   (although generally, for now these are just place-holders, so there
   isn't much emphasis on them now).

 - The comments on each function have been brought up above the
   function. This is the general convention in Gnuastro, in all languages
   ;-).

 - To check if a given argument is a FITS file, a new function called
   '_gnuastro_autocomplete_file_is_fits' is defined. Instead of relying on
   suffixes (which are many!), this will just use the 'astfits' program.

 - A new function called '_gnuastro_autocomplete_option_value' is defined
   to return the value of an option, given its long and short option
   names. It will account for all the scenarios (except for one cases that
   is mentioned as a TODO.

 - The new '_gnuastro_autocomplete_plaintext_is_table' function will return
   0 (success) if its argument is a plain text table that can be parsed in
   Gnuastro.

 - The '_gnuastro_autocomplete_last_table' function replaced the old
   '_gnuastro_autocomplete_get_fits_name' function. It will parse the
   arguments in reverse, and return the last acceptable table. Note that
   FITS files can also be images, and tables can also be in plain-text.

 - The long pipes in some of the options were broken into multiple lines.

 - For 'PROG_NAME', probably '$1' is better than manually setting it. The
   user may use a manual copy of the problem that is not installed in any
   standard location for example. It is better to rely on the user's actual
   executable.

Pedram, once you can guide me on the fix in the '--information' option,
I'll continue with inspecting the code, thanks a lot ;-).