bin/table/completion.bash: Fix '--information'
Until now, if the user asked for information of a fits file, the
autocompletion would still print other options as suggestions. With this
commit, autocompletion checks if the '--information' option is invoked
while also a valid fits file has been specified. In this case, the
autocompletion will cease printing other suggestions, and keep telling
the user that they can just press ENTER and move on.

FIX: This will fix the problem mentioned in the commit below:
e18d46e476dab290087ca7bcb1dd8f096f59923e

The error showed up because there was no variable $NAME, replaced it with
$PROG_NAME and the error '-bash: command : not found' went away.