/bin/table/completion.bash: consider short opt '-i'
The '--information' long and short option '-i' should have the same
autocompletion suggestions. With this commit, the autocompletion will
cease providing suggestions when '-i' or '--information' option is
selected while the fits file is specified.

I could have used the bash extended glob: @(pattern1|pattern2) in the
pattern matching binary operator '=~' inside the '[[ ... ]]' keyword.
But I refrained to do so, since we try not to use more modern operators
to keep the backward compatibility (for Bash older than 3.2 or so).