/bin/table/completion.bash: Refactor code
Until now, I used a case inside an 'if' statement. The syntax looked
pretty confusing and redundant. In this commit, the completion script
will scan the entire prompt for '-i' or '--information' option. If it
finds any instances, it will check if there is a valid fits file
specified. If both cases are TRUE, the completion will print a message
and let the user know that s/he can press ENTER safely.

Also, I declared a new _gnuastro_autocomplete_print_message function.
This function will take only one argument, and pretty prints it. Then,
it will reprint the user prompt with a '$' at line beginning for a
natural and familiar experience.

Last but not least, I used 'grep' to scan the prompt. This is POSIX
compatible and easier to implement when there is more than one pattern
to check.