Auto-completion: Minor polishing of the warning message
Until now, the warning message for the '--information' option was a single
string that was repeated in two places (adding a check for it in general
was a good idea Pedram!). However, based on previous experience, this can
cause bugs if we later edit one of them and forget to the edit the other!

With this commit a new 'infowarning' variable is defined at the top of the
function and message is put there. That variable can now be used anywhere
in the code, and all the time it will have the same value.

Other changes with this commit:

 - When the 'last_table' variable is not empty, the existance of its file
   has already been checked. So we should just check if its an empty string
   or not. Don't forget that checking the existance of a file is
   computationally expensive (requires Bash to leave its internal
   evironment and go to the operating system!).

 - In my case the previous '$' wasn't enough to print the prompt. The new
   line simply started with a '$' sign, not my standard prompt. So I added
   '${PS1@P}'. As mentioned above it, there is a portability issue here
   with older Bash systems. We should find a more portable solution.

 - I removed the extra '-----' lines: they are distracting and waste a lot
   of valuable terminal lines ;-)!

[To help in future reference to commits related to auto-completion, let's
start the commit title with 'Auto-completion'.]