Auto-completion: Smart list fits files, fix var
Until now, the function responsible for listing the fits files only
checked if the file extension ended with 'fits'. This is not a rebust
test, because other fits files -- that might have different extensions
or no extensions at all -- would be left out.

With this commit, I used 'astfits' program return value to check their
validity regardless of their name or extension. However, this has a bit
of overhead and slows response. We can ignore it if it is not that
significant, and will not ruin user experience.

Also, I replaced a hard-coded value with intended $inputfile variable.